misc/the_void/challenge/app/requirements.txt
flask
pillow
pytesseract
apscheduler